Code refactoring
Moved API functions to the new class, moved singleton functions to the Singleton trait
