+++ importer-from-maxsite.php
@@ -6,7 +6,7 @@
 Plugin Name: Importer From MaxSite
 Plugin URI: https://github.com/zahardoc/importer-from-maxsite
 Description: Plugin Importer From MaxSite provides easy and fast way to move your data from MaxSite CMS to the WordPress.
-Version: 1.3
+Version: 1.4
 Author: Sergey Zaharchenko <zaharchenko.dev@gmail.com>
 Author URI: https://github.com/zahardoc
 License: GPLv3
@@ -36,11 +36,13 @@ define( 'IFM_PLUGIN_DIR', plugin_dir_path( __FILE__ ) );
 define( 'IFM_PLUGIN_BASENAME', plugin_basename(__FILE__));
 define( 'IFM_PLUGIN_URL', plugins_url('', __FILE__));
 define( 'IFM_TEXT_DOMAIN', 'importer-from-maxsite');
-define( 'IFM_ASSETS_VERSION', 1.3);
+define( 'IFM_ASSETS_VERSION', 1.4);
 
 load_textdomain(IFM_TEXT_DOMAIN, IFM_PLUGIN_DIR . '/lang/ifm-' . get_locale() . '.mo');
 
+require_once __DIR__ . '/app/trait-singleton.php';
 require_once __DIR__ . '/app/class-page-controller.php';
+require_once __DIR__ . '/app/class-api.php';
 require_once __DIR__ . '/app/class-importer.php';
 
 \Importer_From_Maxsite\Page_Controller::instance()->init();
